This revision adds an example of a new Cucumber feature that fails to a test case that creates a new movie. The problem is with a movie path that is not found. The error message is as follows:
(::) failed steps (::)

Can't find mapping from "the Create New Movie page" to a path.
Now, go and add a mapping in /u1/workspaces/hw2_saas_rottenpotatoes/features/support/paths.rb (RuntimeError)
./features/support/paths.rb:32:in `rescue in path_to'
./features/support/paths.rb:26:in `path_to'
./features/step_definitions/web_steps.rb:235:in `/^(?:|I )should be on (.+)$/'
features/add_movie.feature:6:in `Then I should be on the Create New Movie page'

In order to solve it, as the failure indicates:

from "the Create New Movie page" to a path

Looking at the paths.rb, the cases look for a mapping between the given page name
to a page. In this case, we create a new mapping in the paths file.

* add_movie.feature
- Feature found online to add a new movie manually from the page.

* paths.rb
- Adding a new mapping from the create new page.