his revision updates the Gemfile to add the dependency to Rails autotest, a tool used to run test cases automatically when one changes any test or spec on a running machine. It was briefly described during the presentation of slides 5.4 of the SaaS class from Berkeley.
In order to run, open a different shell under the rails app root directory and run
the command "autotest". The shell will be running and any udpates to test files will
be triggered and the tests run.

The other change on this revision was related to the presentation of slides of chapter
5.4 related to adding the implementation of the test rspec tests to verify that
the calls to the controller tests are attempting to call the model class method.

rspec spec/controllers/out_first_controller_spec.rb

or run

autotest

Another part of the setup described in the class' presentation was to touch on
the search results view file.

* movies_controller.rb
- Adding the call to a non-existing method to the class Movie. This is part of
a seam to be tested.

* search_tmdb.html.haml
- Adding the touched file to make the test pass.

* Gemfile
- Adding the autotest support
- Moving the cucumber related gem to the test only environment.

* Gemfile
- Updated after changing Gemfile