This revision adds the gem "factory_girl" as the default standard for the replacement of fixtures of the Rails tests. This is part of the chapter 5.5 of the SaaS class from Berkeley http://www.scribd.com/doc/86993368/Slides-Chapter5-TDD.
The problem with fixtures is that it introduces dependency to the fixture data,
as it will be executed before every test case.

The solution to that is to use specific data from factories, in this case
using Factory Girl. It helps keep the tests independent by being unnaffected by
the presence of objects that the test does not need. However, it is hard when
there are complex object relationships.

* Gemfile
- Adding the gem 'factory girl' that can be used as a factory to classes.

* Gemfile.lock
- Changed as per the above-mentioned change.

* spec/factories/movies.rb
- Adding the factory girl for movies.